refactor: useLine과 animationController로 분리
