+++ src/components/useAnimationController/useAnimationController.tsx
@@ -1,64 +1,152 @@
 import { useEffect, useRef, useState } from 'react'
+import { MapInstance } from 'react-map-gl'
 
-const useAnimationController = (
-    onUpdate: (deltaTime: number, additionalParam: any) => void,
-    onStart: () => void,
-    onStop: () => void,
-    duration: number) => {
-    const [isPlaying, setIsPlaying] = useState<boolean>(false)
-    const [isPaused, setIsPaused] = useState<boolean>(false)
-    const requestRef = useRef<number | null>(null)
-    const previousTimeRef = useRef<number | null>(null)
+import { FeatureCollection } from '../useLine'
 
-    const play = () => {
-        if (!isPlaying) {
-            setIsPlaying(true)
-            setIsPaused(false)
-            onStart()
-        }
+const useMapAnimationController = ({
+  duration,
+  line,
+  map,
+}: {
+  map: MapInstance | null
+  duration: number
+  line: FeatureCollection | null
+}) => {
+  const [isPlaying, setIsPlaying] = useState<boolean>(false)
+  const [isPaused, setIsPaused] = useState<boolean>(false)
+  const requestRef = useRef<null | number>(null)
+  const previousTimeRef = useRef<null | number>(null)
+  const [currentFrame, setCurrentFrame] = useState<number>(0)
+
+  const draw = (map: any, pos: any) => {
+    if (map?.getSource('point')) {
+      const origin = pos
+      const point = {
+        type: 'FeatureCollection',
+        features: [
+          {
+            type: 'Feature',
+            properties: {},
+            geometry: {
+              type: 'Point',
+              coordinates: origin,
+            },
+          },
+        ],
+      }
+      map.getSource('point').setData(point)
     }
-    const pause = () => {
-        if (isPlaying) {
-            setIsPaused(true)
-            setIsPlaying(false)
-        }
-    }
-    const stop = () => {
-        setIsPlaying(false)
-        setIsPaused(false)
-        if (requestRef.current) {
-            cancelAnimationFrame(requestRef.current)
-            requestRef.current = null
-        }
-        previousTimeRef.current = null
-        onStop()
-    }
+  }
 
-    const animate = (time: number) => {
-        if (previousTimeRef.current !== null) {
-            onUpdate(requestRef.current!, { speed: 1 })
-            console.log(requestRef.current!, duration)
-        }
-        previousTimeRef.current = time
-        if (!isPaused) {
-            requestRef.current = requestAnimationFrame(animate)
-        }
-        if (requestRef.current! >= duration) {
-            cancelAnimationFrame(requestRef.current!)
-        }
+  const animateAtoB = async (
+    map: any,
+    line: any | FeatureCollection /* todo: remove null */
+  ) => {
+    const origin = line?.features[0].geometry.coordinates[0]
+    const point = {
+      type: 'FeatureCollection',
+      features: [
+        {
+          type: 'Feature',
+          properties: {},
+          geometry: {
+            type: 'Point',
+            coordinates: origin,
+          },
+        },
+      ],
     }
 
-    useEffect(() => {
-        if (isPlaying && !isPaused)
-            requestRef.current = requestAnimationFrame(animate)
+    map.loadImage('/airplane.png', (error: Error, image: HTMLIFrameElement) => {
+      if (error) {
+        throw error
+      }
 
-        return () => {
-            if (requestRef.current)
-                cancelAnimationFrame(requestRef.current)
-        }
-    }, [isPlaying, isPaused])
+      !map.hasImage('airplane-icon') && map.addImage('airplane-icon', image)
 
-    return { play, pause, stop, isPlaying, isPaused }
+      !map.getSource('point') &&
+        map.addSource('point', { type: 'geojson', data: point })
+
+      !map.getLayer('points') &&
+        map.addLayer({
+          id: 'points',
+          type: 'symbol',
+          source: 'point',
+          layout: { 'icon-image': 'airplane-icon', 'icon-size': 0.25 },
+        })
+    })
+  }
+
+  const handleUpdate = (deltaTime: number, { speed }: any) => {
+    console.log(deltaTime, speed)
+    setCurrentFrame((prevFrame) => {
+      const nextFrame = prevFrame + 1
+
+      if (line?.features[0].geometry.coordinates[nextFrame])
+        draw(map, line?.features[0].geometry.coordinates[nextFrame])
+
+      return nextFrame
+    })
+  }
+
+  const handleStart = () => {
+    animateAtoB(map, line)
+  }
+
+  const handleStop = () => {
+    setCurrentFrame(0)
+    if (line?.features[0].geometry.coordinates[0])
+      draw(map, line?.features[0].geometry.coordinates[0])
+  }
+
+  const play = () => {
+    if (!isPlaying) {
+      setIsPlaying(true)
+      setIsPaused(false)
+      handleStart()
+    }
+  }
+  const pause = () => {
+    if (isPlaying) {
+      setIsPaused(true)
+      setIsPlaying(false)
+    }
+  }
+  const stop = () => {
+    setIsPlaying(false)
+    setIsPaused(false)
+    if (requestRef.current) {
+      cancelAnimationFrame(requestRef.current)
+      requestRef.current = null
+    }
+    previousTimeRef.current = null
+    handleStop()
+  }
+
+  const animate = (time: number) => {
+    if (previousTimeRef.current !== null) {
+      handleUpdate(requestRef.current!, { speed: 1 })
+      console.log(requestRef.current!, duration)
+    }
+    previousTimeRef.current = time
+    if (!isPaused) {
+      requestRef.current = requestAnimationFrame(animate)
+    }
+    if (requestRef.current! >= duration) {
+      cancelAnimationFrame(requestRef.current!)
+    }
+  }
+
+  useEffect(() => {
+    if (isPlaying && !isPaused)
+      requestRef.current = requestAnimationFrame(animate)
+
+    return () => {
+      if (requestRef.current) cancelAnimationFrame(requestRef.current)
+    }
+  }, [isPlaying, isPaused])
+
+  return { play, pause, stop, isPlaying, isPaused, currentFrame }
 }
 
-export default useAnimationController
+export default useMapAnimationController
