+++ src/pages/FlightOnMap.tsx
@@ -1,8 +1,9 @@
 import React from 'react'
+
 import FlightMap from '../components/VMap/FlightMap'
 
-const FlightOnMap: React.FC = ({ }) => {
-    return (<FlightMap />)
+const FlightOnMap: React.FC = ({}) => {
+  return <FlightMap />
 }
 
 export default FlightOnMap
